api/serializers.py
from django.contrib.auth import get_user_model
from Movies.models import Movie
from rest_framework import serializers

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = get_user_model()
        fields = ['username']
    
        
class MovieSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Movie
        fields = ['title','poster_path','poster_back','overview']