requirements.txt
awscli==1.16.159
awsebcli==3.15.1
beautifulsoup4==4.7.1
blessed==1.15.0
botocore==1.12.149
bs4==0.0.1
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.8
django-bootstrap4==0.0.8
django-filter==2.1.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-seed==0.1.9
djangorestframework==3.9.4
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Faker==1.0.6
Flask==1.0.2
future==0.16.0
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
Markdown==3.1
MarkupSafe==1.1.0
openapi-codec==1.3.2
pathspec==0.5.9
psycopg2==2.8.2
pyasn1==0.4.5
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.20.1
rsa==3.4.2
s3transfer==0.2.0
semantic-version==2.5.0
simplejson==3.16.0
six==1.11.0
soupsieve==1.9.1
termcolor==1.1.0
text-unidecode==1.2
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.14.1
whitenoise==4.1.2
